Move ContextMenu out into its own file.
